Update _VHclip
Consider moving this instruction "<!--Change Vircuh to match your repository-->" below line 142 (<?xml version="1.0"?>) because it appears above it in the clip. Its current position leads to an error when trying to parse and upload the file. I know that when we set up the clips for new repositories we would likely delete this line, but it could cause problems for someone in the future. Thanks.